Ditch arithmetic expression & clean up expression syntax a bit
