+++ argot/src/argot/syntax/mod.rs
@@ -15,10 +15,10 @@ UnaryOperator:  -
 */
 
 pub mod argument_list;
-pub mod arithmetic_expression;
 pub mod atom;
 pub mod atom_expr;
 pub mod block;
+pub mod call;
 pub mod common;
 pub mod expression;
 pub mod factor;
@@ -34,10 +34,10 @@ pub mod var_decl;
 
 pub mod types {
     pub use super::argument_list::ArgumentList;
-    pub use super::arithmetic_expression::ArithmeticExpression;
     pub use super::atom::Atom;
     pub use super::atom_expr::AtomicExpression;
     pub use super::block::Block;
+    pub use super::call::FunctionCall;
     pub use super::expression::Expression;
     pub use super::factor::Factor;
     pub use super::function::FunctionDeclaration;
@@ -50,10 +50,8 @@ pub mod types {
     pub use super::var_decl::{VariableAssignment, VariableDeclaration};
 }
 
-pub use arithmetic_expression::ArithmeticExpression;
 pub use atom::Atom;
 pub use block::Block;
-pub use expression::Expression;
 pub use factor::Factor;
 pub use function::FunctionDeclaration;
 pub use operator::{FactorOperator, TermOperator, UnaryOperator};
