+++ argot/src/argot/visitor.rs
@@ -7,7 +7,6 @@ pub trait Visitable {
 pub trait Visitor {
     type Result;
 
-    fn visit_arithmetic_expression(&mut self, v: &mut ArithmeticExpression) -> Self::Result;
     fn visit_factor(&mut self, v: &mut Factor) -> Self::Result;
     fn visit_factor_operator(&mut self, v: &mut FactorOperator) -> Self::Result;
     fn visit_term(&mut self, v: &mut Term) -> Self::Result;
@@ -22,4 +21,5 @@ pub trait Visitor {
     fn visit_atom(&mut self, v: &mut Atom) -> Self::Result;
     fn visit_block(&mut self, v: &mut Block) -> Self::Result;
     fn visit_variable_assignment(&mut self, v: &mut VariableAssignment) -> Self::Result;
+    fn visit_function_call(&mut self, v: &mut FunctionCall) -> Self::Result;
 }
